Sendgrid verification emails implemented
Implemented the feature of sending verification emails when registering a new user.
Created login verification for whether the email of the current user is verified.
Created endpoint for resending the verification email.
Changed swagger documentation.
Repository users created.
Changed user model so that it included verification token and "verified" field.
Constants for my personal ease of use added.
Controller for verifying the email and resending it was created.